Parsing source elements to math
