+++ lib/iev/termbase/converter.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module IEV
+  module Termbase
+    module Converter
+      def self.mathml_to_asciimath(input)
+        IEV::Termbase::Converter::MathmlToAsciimath.convert(input)
+      end
+    end
+  end
+end
